Move vision registration to vision.ts
