+++ src/server/plugins/engine/index.ts
@@ -1,7 +1,7 @@
 import { type Environment } from 'nunjucks'
 
 import { engine } from '~/src/server/plugins/engine/helpers.js'
-import { plugin } from '~/src/server/plugins/engine/plugin.js'
+import { plugin } from './vision.js'
 import { type FilterFunction } from '~/src/server/plugins/engine/types.js'
 import {
   checkComponentTemplates,
